third_party/blink/web_tests/wpt_internal/speculation-rules/prefetch/link-selection-heuristics.https.html

<!DOCTYPE html>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="/speculation-rules/prefetch/resources/utils.sub.js"></script>

<meta name="variant" content="?event_type=none">
<meta name="variant" content="?event_type=pointer-down">
<meta name="variant" content="?event_type=pointer-hover">
<script>
  // This is split across different test variants due to the test timeouts.
  let {event_type} = Object.fromEntries(new URLSearchParams(location.search));
  promise_test(async t => {
    assert_implements(
        HTMLScriptElement.supports('speculationrules'),
        'Speculation Rules not supported');

    insertSpeculationRules({
      prefetch: [
        {
          source: 'document',
          where: {href_matches: '/*\\?*'}
        }
      ]
    });

    function insertLink(id, url) {
      let link = document.createElement('a');
      link.id = id;
      link.href = url;
      link.innerHTML = '<div id="testDiv" style="-webkit-user-select: none; border:1px solid red">Click inside this div</div>';
      document.body.appendChild(link);
      return link;
    }

    document.body = document.createElement("body");
    let url = getPrefetchUrlList(1)[0];
    testLink = insertLink('testLink', url);
    dummyLink = insertLink('dummyLink', '');

    let x = 1;
    let y = 1;
    if (event_type === "pointer-down") {
      await new test_driver.Actions()
      .pointerMove(x, y, {origin:testLink})
      .pointerDown()
      .pointerMove(x, y, {origin:dummyLink})
      .send();
    } else if (event_type === "pointer-hover") {
      const pause_time_in_ms = 500;
      await new test_driver.Actions()
      .pointerMove(x, y, {origin:testLink})
      .pause(pause_time_in_ms)
      .send();
    } else if (event_type === "none") {
      await new test_driver.Actions()
      .pointerMove(x, y, {origin:testLink})
      .pointerMove(x, y, {origin:dummyLink})
      .send();
    }

    await new Promise(resolve => t.step_timeout(resolve, 2000));

    if (event_type !== 'none') {
      assert_equals(await isUrlPrefetched(url), 1, "Should be prefetched by heuristics.");
    } else {
      assert_equals(await isUrlPrefetched(url), 0, "Should not be prefetched by heuristics.");
    }
  }, `test on pointer event link selection heuristics.`);
</script>
